back_end/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.2
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
docx2txt==0.8
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.104.1
fuzzywuzzy==0.18.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
lxml==6.0.1
openpyxl==3.1.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.13
pydantic_core==2.33.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.24.0
